requirements.txt
# Stratobet Backend - Production Requirements

# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.0
httpx==0.25.2

# ML
joblib==1.3.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
